requirements.txt
ccxt
structlog
python-json-logger
PyYAML
TA-lib
pandas
tushare